+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_model_service.py
@@ -23,6 +23,7 @@ class ModelService(DaraModel):
         provider_settings: main_models.ProviderSettings = None,
         status: str = None,
         status_reason: str = None,
+        workspace_id: str = None,
     ):
         self.created_at = created_at
         self.credential_name = credential_name
@@ -37,6 +38,7 @@ class ModelService(DaraModel):
         self.provider_settings = provider_settings
         self.status = status
         self.status_reason = status_reason
+        self.workspace_id = workspace_id
 
     def validate(self):
         if self.model_info_configs:
@@ -94,6 +96,9 @@ class ModelService(DaraModel):
         if self.status_reason is not None:
             result['statusReason'] = self.status_reason
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -141,6 +146,9 @@ class ModelService(DaraModel):
 
         if m.get('statusReason') is not None:
             self.status_reason = m.get('statusReason')
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
 
         return self