+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_update_model_service_input.py
@@ -17,6 +17,7 @@ class UpdateModelServiceInput(DaraModel):
         provider_settings: main_models.ProviderSettings = None,
         status: str = None,
         status_reason: str = None,
+        workspace_id: str = None,
     ):
         self.credential_name = credential_name
         self.description = description
@@ -25,6 +26,7 @@ class UpdateModelServiceInput(DaraModel):
         self.provider_settings = provider_settings
         self.status = status
         self.status_reason = status_reason
+        self.workspace_id = workspace_id
 
     def validate(self):
         if self.model_info_configs:
@@ -64,6 +66,9 @@ class UpdateModelServiceInput(DaraModel):
         if self.status_reason is not None:
             result['statusReason'] = self.status_reason
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -93,6 +98,9 @@ class UpdateModelServiceInput(DaraModel):
 
         if m.get('statusReason') is not None:
             self.status_reason = m.get('statusReason')
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
 
         return self